dma_heap: allow building dma_heap framework as a module
DMA heap is a thin framework on top of dmabuf to help implement
function specific DMA heaps with different allocation and mapping
methods like CMA and System heaps.

Build this as a module can help OS distributions to release a general
kernel with this function.

Or just export symbols? Since some DMA heaps depending on this framework
can't be compiled as modules. TODO

Signed-off-by: Woody Zhang <woodyzhang666@gmail.com>