debugobjects: use hlist_bl to save memory
This saves about 512KB memory for hash slot spinlocks.
bit spinlock seems to be much slower than normal spinlock.
bit spinlock and hlist_bl seem not be used widely. And a lot of API
variants are not added, like -irqsave version of bitlock, Movement of
hlist_bl.

So this work may not be accepted by maintainers.

Signed-off-by: Woody Zhang <woodyzhang666@gmail.com>